+++ src/compas_model/algorithms/__init__.py
@@ -1,17 +1,50 @@
-from .collisions import is_aabb_aabb_collision
-from .collisions import is_box_box_collision
-from .collisions import is_face_to_face_collision
+from .intersections import is_intersection_line_aabb
+from .intersections import is_intersection_line_box
+from .intersections import is_intersection_ray_aabb
+from .intersections import is_intersection_ray_box
+from .intersections import is_intersection_segment_aabb
+from .intersections import is_intersection_segment_box
+from .intersections import is_intersection_box_box
 
-from .collisions import get_collision_pairs  # rename to model_collisions
+from .intersections import intersection_ray_triangle
+
+from .intersections import intersections_line_aabb
+from .intersections import intersections_line_box
+from .intersections import intersections_ray_aabb
+from .intersections import intersections_ray_box
+
+from .minkowski2 import minkowski_sum_xy
+from .minkowski2 import minkowski_difference_xy
+
+from .gjk2 import is_collision_poly_poly_xy
+
+from .bvh import AABBNode
+from .bvh import OBBNode
+from .bvh import BVH
+
 from .interfaces import model_interfaces
 from .overlaps import model_overlaps
 
 
 __all__ = [
-    "is_aabb_aabb_collision",
-    "is_box_box_collision",
-    "is_face_to_face_collision",
-    "get_collision_pairs",
+    "AABBNode",
+    "OBBNode",
+    "BVH",
+    "intersection_ray_triangle",
+    "intersections_line_aabb",
+    "intersections_line_box",
+    "intersections_ray_aabb",
+    "intersections_ray_box",
+    "is_collision_poly_poly_xy",
+    "is_intersection_box_box",
+    "is_intersection_line_aabb",
+    "is_intersection_line_box",
+    "is_intersection_ray_aabb",
+    "is_intersection_ray_box",
+    "is_intersection_segment_aabb",
+    "is_intersection_segment_box",
+    "minkowski_difference_xy",
+    "minkowski_sum_xy",
     "model_interfaces",
     "model_overlaps",
 ]