+++ src/compas_model/algorithms/minkowski2.py
@@ -3,55 +3,133 @@ from math import atan2
 from compas.geometry import Point
 from compas.geometry import Polygon
 from compas.geometry import Vector
-from compas.geometry import centroid_points_xy
-from compas.geometry import distance_point_line_sqrd_xy
-from compas.geometry import dot_vectors_xy
-from compas.itertools import pairwise
+from compas.geometry import centroid_points
+
+# =============================================================================
+# Helpers
+# =============================================================================
 
 
-def cross(u: Vector, v: Vector) -> float:
+def det(u: Vector, v: Vector) -> float:
+    """Compute the determinant of the 2x2 matrix formed by the XY components of the vectors u and v.
+
+    Parameters
+    ----------
+    u : :class:`compas.geometry.Vector`
+        The first vector.
+    v : :class:`compas.geometry.Vector`
+        The second vector.
+
+    Returns
+    -------
+    float
+
+    """
     return u[0] * v[1] - u[1] * v[0]
 
 
-def bottomleft(points) -> int:
+def bottomleft(points: list[Point]) -> int:
+    """Identify the point with the smalles y coordinate, and the smallest x coordinate.
+
+    Parameters
+    ----------
+    points : list[:class:`compas.geometry.Point`]
+        A list of points.
+
+    Returns
+    -------
+    int
+        The index of the point.
+
+    """
     return sorted(enumerate(points), key=lambda x: (x[1][1], x[1][0]))[0][0]
 
 
-def sort_ccw(points: list[Point]):
-    cx, cy, cz = centroid_points_xy(points)
+def sort_ccw(points: list[Point]) -> list[Point]:
+    """Sort the points in CCW direction using the polar angle wrt their centroid.
+
+    Parameters
+    ----------
+    points : list[:class:`compas.geometry.Point`]
+        A list of points.
+
+    Returns
+    -------
+    list[:class:`compas.geometry.Point`]
+        The sorted points.
+
+    """
+    cx, cy, _ = centroid_points(points)
     indices, points = zip(*sorted(enumerate(points), key=lambda x: atan2(x[1][1] - cy, x[1][0] - cx)))
     return [points[i] for i in indices]
 
 
-def sort_bottomleft(points: list[Point]) -> list[Point]:
+def reorder_bottomleft(points: list[Point]) -> list[Point]:
+    """Reorder the points to start with the one on the bottom left.
+
+    Parameters
+    ----------
+    points : list[:class:`compas.geometry.Point`]
+        A list of points.
+
+    Returns
+    -------
+    list[:class:`compas.geometry.Point`]
+        The reordered points.
+
+    """
     index = bottomleft(points)
     return points[index:] + points[:index]
 
 
-def support_point(A: list[Vector], direction: Vector) -> Vector:
-    return sorted(A, key=lambda point: dot_vectors_xy(point, direction))[-1]
+# =============================================================================
+# Minkowski
+# =============================================================================
 
 
-def support(A: list[Vector], B: list[Vector], direction: Vector) -> Vector:
-    return support_point(A, direction) - support_point(B, direction * -1)
+def minkowski_sum_xy(A: Polygon, B: Polygon) -> Polygon:
+    """Compute the Minkowski sum of two polygons.
 
+    Parameters
+    ----------
+    A : :class:`compas.geometry.Polygon`
+        The first polygon.
+    B : :class:`compas.geometry.Polygon`
+        The second polygon.
 
-def closest_edge_to_origin(abc) -> tuple[Point, Point]:
-    return sorted(pairwise(abc + abc[:1]), key=lambda line: distance_point_line_sqrd_xy([0, 0, 0], line))[0]
+    Returns
+    -------
+    :class:`compas.geometry.Polygon`
+        The polygon representing the sum.
 
+    Warnings
+    --------
+    Currently only convex polygons are supported.
 
-def minksum2(A: Polygon, B: Polygon) -> Polygon:
+    See Also
+    --------
+    minkdiff2
+
+    Notes
+    -----
+    ...
+
+    References
+    ----------
+    ...
+
+    """
     points = []
 
-    A = sort_bottomleft(sort_ccw(A))
-    B = sort_bottomleft(sort_ccw(B))
+    A = reorder_bottomleft(sort_ccw(A))
+    B = reorder_bottomleft(sort_ccw(B))
 
     i, j = 0, 0
     la, lb = len(A), len(B)
 
     while i < la or j < lb:
         points.append(A[i % la] + B[j % lb])
-        d = cross(A[(i + 1) % la] - A[i % la], B[(j + 1) % lb] - B[j % lb])
+        d = det(A[(i + 1) % la] - A[i % la], B[(j + 1) % lb] - B[j % lb])
         if d >= 0:
             i += 1
         if d <= 0:
@@ -60,5 +138,36 @@ def minksum2(A: Polygon, B: Polygon) -> Polygon:
     return Polygon(points)
 
 
-def minkdiff2(A: Polygon, B: Polygon) -> Polygon:
-    return minksum2(A, [Point(-x, -y, -z) for x, y, z in B])
+def minkowski_difference_xy(A: Polygon, B: Polygon) -> Polygon:
+    """Compute the Minkowski sum of polygon A and inverted polygon B.
+
+    Parameters
+    ----------
+    A : :class:`compas.geometry.Polygon`
+        The first polygon.
+    B : :class:`compas.geometry.Polygon`
+        The second polygon.
+
+    Returns
+    -------
+    :class:`compas.geometry.Polygon`
+        The polygon representing the difference as the sum of A and -B.
+
+    Warnings
+    --------
+    Currently only convex polygons are supported.
+
+    See Also
+    --------
+    minksum2
+
+    Notes
+    -----
+    ...
+
+    References
+    ----------
+    ...
+
+    """
+    return minkowski_sum_xy(A, [Point(-x, -y, -z) for x, y, z in B])